docker/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# HTTP Client
httpx==0.27.2

# Database
psycopg2-binary==2.9.10

# Environment Variables
python-dotenv==1.0.1
